feat: 워크북 특정 가게의 리뷰목록 보기 API 구현

+++ src/main/java/umc/domain/review/dto/req/ReviewReqDTO.java
@@ -14,7 +14,7 @@ public class ReviewReqDTO {
             @NotNull
             @Min(value = 0, message = "별점은 0점 이상이어야 합니다.")
             @Max(value = 5, message = "별점은 5점 이하여야 합니다.")
-            Integer rating,
+            Float rating,
 
             String content
     ){}
